requirements.txt
paho-mqtt==1.3.1
pyTelegramBotAPI==3.6.3
requests==2.7.0
six==1.11.0
telebot==0.0.3
